% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bnkpg_df}
\alias{bnkpg_df}
\title{Brunei kampong data}
\format{
A tibble containing kampong, mukim, district information. All rows
  correspond to an area on the map. Useful to combine with other unit-level
  information for plotting.
}
\usage{
bnkpg_df
}
\description{
Brunei kampong data
}
\details{
Brunei kampong data
}
\keyword{datasets}
