% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruneimap-package.R
\docType{package}
\name{bruneimap-package}
\alias{bruneimap-package}
\title{bruneimap: Maps and Spatial Data of Brunei}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides spatial data for mapping Brunei, including boundaries for districts, mukims, and kampongs, as well as locations of key infrastructure such as masjids, hospitals, clinics, and schools. The package supports researchers, analysts, and developers working with Brunei’s geographic and demographic data, offering a quick and accessible foundation for creating maps and conducting spatial studies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bruneiverse.github.io/bruneimap/}
  \item Report bugs at \url{https://github.com/Bruneiverse/bruneimap/issues}
}

}
\author{
\strong{Maintainer}: Haziq Jamil \email{haziq.jamil@gmail.com} (\href{https://orcid.org/0000-0003-3298-1010}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Hafeezul Raziq \email{hafeezul.rzq@gmail.com} [contributor]
  \item Alvin Bong [contributor]
  \item Aniq Najwa Ariffin [contributor]
  \item Danish Ikhwan Mohamad Zainin [contributor]
  \item Muhammad Rayme Hijazi Jassrie [contributor]
  \item Angela Chan [contributor]
  \item Irdina Batrisyia Norharddyman [contributor]
  \item Nursyafiqah Mohammad Yusuf [contributor]
  \item Ahmad Syafi Mohammad Hilmy [contributor]
  \item Mohammad Syahir Deedat Al Jufri [contributor]
  \item Nadi.BN \email{info@aiti.gov.bn} (CC BY 4.0 <https://www.nadi.bn/>) [copyright holder]
}

}
\keyword{internal}
