% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\docType{data}
\name{census2021}
\alias{census2021}
\title{Brunei census data 2021}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 365 rows and 11 columns.
}
\usage{
census2021
}
\description{
Brunei census data 2021
}
\details{
A tibble containing the population, split into male/female or
Bruneian/PR/foreigners for each kampong. Also contains number of households
and occupied living quarters.
}
\examples{
library(tidyverse)
census2021 |>
  summarise(
    population = sum(population),
    .by = district
  )
}
\references{
DEPS. “The Population and Housing Census Report (BPP) 2021:
Demographic, Household and Housing Characteristics.” Department of Economic
Planning and Statistics, Ministry of Finance and Economy, Brunei
Darussalam, October 2022. URL:
\url{https://deps.mofe.gov.bn/SitePages/Population.aspx}
}
\keyword{datasets}
