% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestreserve.R
\docType{data}
\name{fr_sf}
\alias{fr_sf}
\title{Forest reserve areas in Brunei}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 20 rows and 6 columns.
}
\usage{
fr_sf
}
\description{
Forest reserve areas in Brunei
}
\examples{
library(ggplot2)
ggplot(fr_sf) +
  geom_sf()
}
\references{
Forestry Dept., . “The 5th National Report to the Convention on Biological Diversity.” Bandar Seri Begawan, Brunei: The Forestry Department, Ministry of Industry and Primary Resources, 2014.
}
\keyword{datasets}
