% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthcare.R
\docType{data}
\name{healthcare}
\alias{healthcare}
\title{Hospitals and Clinics (Health Centres) in Brunei}
\format{
A tibble containing the name, type, and lat/long coordinates of
hospitals and clinics in Brunei.
}
\usage{
healthcare
}
\description{
Hospitals and Clinics (Health Centres) in Brunei
}
\examples{
nrow(healthcare)
ncol(healthcare)
}
\keyword{datasets}
