% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masjid.R
\docType{data}
\name{masjid}
\alias{masjid}
\title{Masjid (mosques) in Brunei}
\format{
A tibble containing the name and lat/long coordinates of masjids in
Brunei
}
\source{
\itemize{
\item Ministry of Religious Affairs Annual Statistics 2021. URL: \url{https://www.mora.gov.bn/SiteCollectionDocuments/Statistik/Perangkaan\%20Tahunan\%20KHEU\%202021.pdf}
\item Senarai Masjid di Negara Brunei Darussalam. URL: <https://www.mora.gov.bn/senaraimasjid/forms/viewall.aspx
}
}
\usage{
masjid
}
\description{
Masjid (mosques) in Brunei
}
\examples{
dplyr::glimpse(masjid)
}
\keyword{datasets}
