% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schools.R
\docType{data}
\name{schools}
\alias{schools}
\alias{sch_sf}
\alias{tchr}
\alias{enrolment}
\alias{enrolment_MOE}
\title{Data sets relating to schools in Brunei}
\format{
Either a \code{\link{tibble}} or an \code{\link{sf}} object.

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 252 rows and 14 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 48 rows and 5 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 48 rows and 5 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 84 rows and 5 columns.
}
\usage{
sch_sf

tchr

enrolment

enrolment_MOE
}
\description{
Data sets relating to schools in Brunei
}
\details{
There are four datasets related to schools in Brunei:
\itemize{
\item \code{sch_sf}: A spatial object of schools in Brunei
\item \code{tchr}: A tibble of the number of teachers in schools in Brunei
\item \code{enrolment}: A tibble of the number of students in schools in Brunei
\item \code{enrolment_MOE}: A tibble of the number of students in schools in Brunei
}
}
\keyword{datasets}
