% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\docType{data}
\name{bnkpg_df}
\alias{bnkpg_df}
\title{Brunei kampong data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 438 rows and 4 columns.
}
\usage{
bnkpg_df
}
\description{
Brunei kampong data
}
\details{
A tibble containing kampong, mukim, district information. All rows
correspond to an area on the map. Useful to combine with other unit-level
information for plotting.
}
\keyword{datasets}
