% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruneimap-package.R
\docType{package}
\name{bruneimap-package}
\alias{bruneimap-package}
\title{bruneimap: Maps and Spatial Data of Brunei}
\description{
The bruneimap package provides a rich collection of spatial data in sf format for Brunei, including boundaries for districts, mukims, and kampongs, as well as a silhouette map of the country. Additional datasets include locations of masjids, hospitals, clinics, and schools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Bruneiverse/bruneimap}
  \item \url{https://bruneiverse.github.io/bruneimap/}
  \item Report bugs at \url{https://github.com/Bruneiverse/bruneimap/issues}
}

}
\author{
\strong{Maintainer}: Haziq Jamil \email{haziq.jamil@gmail.com} (\href{https://orcid.org/0000-0003-3298-1010}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Hafeezul Raziq [contributor]
  \item Alvin Bong [contributor]
  \item Aniq Najwa Ariffin [contributor]
  \item Danish Ikhwan Mohamad Zainin [contributor]
  \item Muhammad Rayme Hijazi Jassrie [contributor]
  \item Angela Chan [contributor]
  \item Irdina Batrisyia Norharddyman [contributor]
  \item Nursyafiqah Mohammad Yusuf [contributor]
  \item Ahmad Syafi Mohammad Hilmy [contributor]
  \item Mohammad Syahir Deedat Al Jufri [contributor]
}

}
\keyword{internal}
