% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bruneimap}
\alias{bruneimap}
\alias{dis_sf}
\alias{mkm_sf}
\alias{kpg_sf}
\title{Brunei map}
\format{
An `sf` object
}
\usage{
dis_sf

mkm_sf

kpg_sf
}
\description{
Brunei map
}
\details{
Brunei map
}
\keyword{datasets}
