% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masjid.R
\docType{data}
\name{masjid}
\alias{masjid}
\title{Masjid (mosques) in Brunei}
\format{
A tibble containing the name and lat/long coordinates of masjids in
Brunei
}
\usage{
masjid
}
\description{
Masjid (mosques) in Brunei
}
\examples{
dplyr::glimpse(masjid)
}
\references{
https://www.mora.gov.bn/senaraimasjid/forms/viewall.aspx
}
\keyword{datasets}
